requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
mysqlclient
PyJWT
pytz
sqlparse
